+++ packages/ai/src/tools/classes/index.ts
@@ -1,22 +1,19 @@
-// ASK mode tools
+export { BashEditTool } from './bash-edit';
 export { BashReadTool } from './bash-read';
-export { ListFilesTool } from './list-files';
-export { ReadFileTool } from './read-file';
-export { ScrapeUrlTool } from './scrape-url';
-export { WebSearchTool } from './web-search';
-export { OnlookInstructionsTool } from './onlook-instructions';
-export { ReadStyleGuideTool } from './read-style-guide';
-export { ListBranchesTool } from './list-branches';
+export { CheckErrorsTool } from './check-errors';
+export { FuzzyEditFileTool } from './fuzzy-edit-file';
 export { GlobTool } from './glob';
 export { GrepTool } from './grep';
-
-// BUILD mode tools
+export { ListBranchesTool } from './list-branches';
+export { ListFilesTool } from './list-files';
+export { OnlookInstructionsTool } from './onlook-instructions';
+export { ReadFileTool } from './read-file';
+export { ReadStyleGuideTool } from './read-style-guide';
+export { SandboxTool } from './sandbox';
+export { ScrapeUrlTool } from './scrape-url';
 export { SearchReplaceEditTool } from './search-replace-edit';
 export { SearchReplaceMultiEditFileTool } from './search-replace-multi-edit';
-export { FuzzyEditFileTool } from './fuzzy-edit-file';
-export { WriteFileTool } from './write-file';
-export { BashEditTool } from './bash-edit';
-export { SandboxTool } from './sandbox';
 export { TerminalCommandTool } from './terminal-command';
 export { TypecheckTool } from './typecheck';
-export { CheckErrorsTool } from './check-errors';
+export { WebSearchTool } from './web-search';
+export { WriteFileTool } from './write-file';
