+++ packages/ai/src/tools/classes/list-branches.ts
@@ -1,24 +1,25 @@
+import type { Branch } from '@onlook/models';
 import { Icons } from '@onlook/ui/icons';
+import { jsonClone } from '@onlook/utility/src/clone';
+import type { EditorEngine } from '@onlook/web-client/src/components/store/editor/engine';
 import { z } from 'zod';
-import { ClientTool, type EditorEngine } from '../models/client';
+import { ClientTool } from '../models/client';
 
-export class ListBranchesTool extends ClientTool {
+export class ListBranchesTool implements ClientTool {
     static readonly name = 'list_branches';
     static readonly description = 'List all available branches in the project';
     static readonly parameters = z.object({});
     static readonly icon = Icons.Branch;
 
-    constructor(
-        private handleImpl?: (input: z.infer<typeof ListBranchesTool.parameters>, editorEngine: EditorEngine) => Promise<any>
-    ) {
-        super();
-    }
-
-    async handle(input: z.infer<typeof ListBranchesTool.parameters>, editorEngine: EditorEngine): Promise<any> {
-        if (this.handleImpl) {
-            return this.handleImpl(input, editorEngine);
-        }
-        throw new Error('ListBranchesTool.handle must be implemented by providing handleImpl in constructor');
+    async handle(_params: unknown, editorEngine: EditorEngine): Promise<{
+        branches: Branch[];
+        activeBranchId: string | null;
+    }> {
+        const branches = jsonClone(editorEngine.branches.allBranches)
+        return {
+            branches,
+            activeBranchId: editorEngine.branches.activeBranch?.id || null,
+        };
     }
 
     getLabel(input?: z.infer<typeof ListBranchesTool.parameters>): string {
