+++ packages/ai/src/tools/classes/read-style-guide.ts
@@ -1,24 +1,25 @@
 import { Icons } from '@onlook/ui/icons';
+import type { EditorEngine } from '@onlook/web-client/src/components/store/editor/engine';
 import { z } from 'zod';
-import { ClientTool, type EditorEngine } from '../models/client';
+import { ClientTool } from '../models/client';
 
-export class ReadStyleGuideTool extends ClientTool {
+export class ReadStyleGuideTool implements ClientTool {
     static readonly name = 'read_style_guide';
     static readonly description = 'Read the project style guide and coding conventions';
     static readonly parameters = z.object({});
     static readonly icon = Icons.Brand;
 
-    constructor(
-        private handleImpl?: (input: z.infer<typeof ReadStyleGuideTool.parameters>, editorEngine: EditorEngine) => Promise<any>
-    ) {
-        super();
-    }
-
-    async handle(input: z.infer<typeof ReadStyleGuideTool.parameters>, editorEngine: EditorEngine): Promise<any> {
-        if (this.handleImpl) {
-            return this.handleImpl(input, editorEngine);
+    async handle(_params: unknown, editorEngine: EditorEngine): Promise<{
+        configPath: string;
+        cssPath: string;
+        configContent: string;
+        cssContent: string;
+    }> {
+        const result = await editorEngine.theme.initializeTailwindColorContent();
+        if (!result) {
+            throw new Error('Style guide files not found');
         }
-        throw new Error('ReadStyleGuideTool.handle must be implemented by providing handleImpl in constructor');
+        return result;
     }
 
     getLabel(input?: z.infer<typeof ReadStyleGuideTool.parameters>): string {
