+++ packages/ai/src/tools/classes/search-replace-edit.ts
@@ -1,31 +1,60 @@
 import { Icons } from '@onlook/ui/icons';
+import type { EditorEngine } from '@onlook/web-client/src/components/store/editor/engine';
 import { z } from 'zod';
-import { ClientTool, type EditorEngine } from '../models/client';
+import { ClientTool } from '../models/client';
+import { BRANCH_ID_SCHEMA } from '../shared/type';
 
-export class SearchReplaceEditTool extends ClientTool {
+export class SearchReplaceEditTool implements ClientTool {
     static readonly name = 'search_replace_edit_file';
-    static readonly description = 'Performs exact string replacements in files. The edit will FAIL if `old_string` is not unique in the file. Either provide a larger string with more surrounding context to make it unique or use `replace_all` to change every instance of `old_string`.';
+    static readonly description = 'Performs exact string replacements in files. The edit will FAIL if `old_string` is not unique in the file. Either provide a larger string with more surrounding context to make it unique or use `replace_all` to change every instance of `old_string`.'
     static readonly parameters = z.object({
         file_path: z.string().describe('Absolute path to file'),
         old_string: z.string().describe('Text to replace'),
         new_string: z.string().describe('Replacement text'),
         replace_all: z.boolean().optional().default(false).describe('Replace all occurrences'),
-        branchId: z.string().optional().describe('The branch ID to operate on'),
+        branchId: BRANCH_ID_SCHEMA,
     });
     static readonly icon = Icons.Pencil;
 
-    constructor(
-        private handleImpl?: (input: z.infer<typeof SearchReplaceEditTool.parameters>, editorEngine: EditorEngine) => Promise<any>
-    ) {
-        super();
+    async handle(args: z.infer<typeof SearchReplaceEditTool.parameters>, editorEngine: EditorEngine): Promise<string> {
+        try {
+            const sandbox = editorEngine.branches.getSandboxById(args.branchId);
+            if (!sandbox) {
+                throw new Error(`Sandbox not found for branch ID: ${args.branchId}`);
+            }
+            const file = await sandbox.readFile(args.file_path);
+            if (!file || file.type !== 'text') {
+                throw new Error(`Cannot read file ${args.file_path}: file not found or not text`);
+            }
+
+            let newContent: string;
+            if (args.replace_all) {
+                newContent = file.content.replaceAll(args.old_string, args.new_string);
+            } else {
+                const firstIndex = file.content.indexOf(args.old_string);
+                if (firstIndex === -1) {
+                    throw new Error(`String not found in file: ${args.old_string}`);
+                }
+
+                const secondIndex = file.content.indexOf(args.old_string, firstIndex + args.old_string.length);
+                if (secondIndex !== -1) {
+                    throw new Error(`Multiple occurrences found. Use replace_all=true or provide more context.`);
+                }
+
+                newContent = file.content.replace(args.old_string, args.new_string);
+            }
+
+            const result = await sandbox.writeFile(args.file_path, newContent);
+            if (!result) {
+                throw new Error(`Failed to write file ${args.file_path}`);
+            }
+
+            return `File ${args.file_path} edited successfully`;
+        } catch (error) {
+            throw new Error(`Cannot edit file ${args.file_path}: ${error}`);
+        }
     }
 
-    async handle(input: z.infer<typeof SearchReplaceEditTool.parameters>, editorEngine: EditorEngine): Promise<any> {
-        if (this.handleImpl) {
-            return this.handleImpl(input, editorEngine);
-        }
-        throw new Error('SearchReplaceEditTool.handle must be implemented by providing handleImpl in constructor');
-    }
 
     getLabel(input?: z.infer<typeof SearchReplaceEditTool.parameters>): string {
         if (input?.file_path) {
