+++ packages/ai/src/tools/classes/search-replace-multi-edit.ts
@@ -1,32 +1,90 @@
 import { Icons } from '@onlook/ui/icons';
+import type { EditorEngine } from '@onlook/web-client/src/components/store/editor/engine';
 import { z } from 'zod';
-import { ClientTool, type EditorEngine } from '../models/client';
+import { ClientTool } from '../models/client';
+import { BRANCH_ID_SCHEMA } from '../shared/type';
 
-export class SearchReplaceMultiEditFileTool extends ClientTool {
+export class SearchReplaceMultiEditFileTool implements ClientTool {
     static readonly name = 'search_replace_multi_edit_file';
     static readonly description = 'Perform multiple search and replace operations in a file';
     static readonly parameters = z.object({
-        file_path: z.string().describe('Path to the file to edit'),
-        edits: z.array(z.object({
-            old_string: z.string().describe('Text to search for'),
-            new_string: z.string().describe('Text to replace with'),
-            replace_all: z.boolean().optional().default(false).describe('Replace all occurrences'),
-        })).describe('Array of search and replace operations'),
-        branchId: z.string().optional().describe('The branch ID to operate on'),
+        file_path: z.string().describe('Absolute path to file'),
+        edits: z
+            .array(
+                z.object({
+                    old_string: z.string().describe('Text to replace'),
+                    new_string: z.string().describe('Replacement text'),
+                    replace_all: z
+                        .boolean()
+                        .optional()
+                        .default(false)
+                        .describe('Replace all occurrences'),
+                }),
+            )
+            .describe('Array of edit operations'),
+        branchId: BRANCH_ID_SCHEMA,
     });
     static readonly icon = Icons.Pencil;
 
-    constructor(
-        private handleImpl?: (input: z.infer<typeof SearchReplaceMultiEditFileTool.parameters>, editorEngine: EditorEngine) => Promise<any>
-    ) {
-        super();
-    }
+    async handle(args: z.infer<typeof SearchReplaceMultiEditFileTool.parameters>, editorEngine: EditorEngine): Promise<string> {
+        try {
+            const sandbox = editorEngine.branches.getSandboxById(args.branchId);
+            if (!sandbox) {
+                throw new Error(`Sandbox not found for branch ID: ${args.branchId}`);
+            }
+            const file = await sandbox.readFile(args.file_path);
+            if (!file || file.type !== 'text') {
+                throw new Error(`Cannot read file ${args.file_path}: file not found or not text`);
+            }
 
-    async handle(input: z.infer<typeof SearchReplaceMultiEditFileTool.parameters>, editorEngine: EditorEngine): Promise<any> {
-        if (this.handleImpl) {
-            return this.handleImpl(input, editorEngine);
+            const originalContent = file.content;
+            let content = originalContent;
+
+            // Validate only the first non-replace_all edit against original content
+            // Sequential edits will be validated during application
+            let tempContent = originalContent;
+            for (const edit of args.edits) {
+                if (!edit.replace_all) {
+                    const firstIndex = tempContent.indexOf(edit.old_string);
+                    if (firstIndex === -1) {
+                        throw new Error(`String not found in file: ${edit.old_string}`);
+                    }
+
+                    const secondIndex = tempContent.indexOf(edit.old_string, firstIndex + edit.old_string.length);
+                    if (secondIndex !== -1) {
+                        throw new Error(`Multiple occurrences found for "${edit.old_string}". Use replace_all=true or provide more context.`);
+                    }
+
+                    // Simulate the edit for next validation
+                    tempContent = tempContent.replace(edit.old_string, edit.new_string);
+                } else {
+                    tempContent = tempContent.replaceAll(edit.old_string, edit.new_string);
+                }
+            }
+
+            // Apply edits sequentially in the order provided
+            // Each edit operates on the result of the previous edit
+            for (const edit of args.edits) {
+                if (edit.replace_all) {
+                    content = content.replaceAll(edit.old_string, edit.new_string);
+                } else {
+                    const index = content.indexOf(edit.old_string);
+                    if (index === -1) {
+                        throw new Error(`String not found in file after previous edits: ${edit.old_string}`);
+                    }
+                    content = content.replace(edit.old_string, edit.new_string);
+                }
+            }
+
+            const result = await sandbox.writeFile(args.file_path, content);
+            if (!result) {
+                throw new Error(`Failed to write file ${args.file_path}`);
+            }
+
+            return `File ${args.file_path} edited with ${args.edits.length} changes`;
+        } catch (error) {
+            throw new Error(`Cannot multi-edit file ${args.file_path}: ${(error as Error).message}`);
         }
-        throw new Error('SearchReplaceMultiEditFileTool.handle must be implemented by providing handleImpl in constructor');
     }
 
     getLabel(input?: z.infer<typeof SearchReplaceMultiEditFileTool.parameters>): string {
